feat: add method about user talk agree and collect get

+++ app/creation/service/internal/biz/talk.go
@@ -5,6 +5,7 @@ import (
 	kerrors "github.com/go-kratos/kratos/v2/errors"
 	"github.com/go-kratos/kratos/v2/log"
 	v1 "github.com/the-zion/matrix-core/api/creation/service/v1"
+	"golang.org/x/sync/errgroup"
 )
 
 type TalkRepo interface {
@@ -22,7 +23,10 @@ type TalkRepo interface {
 	GetTalkAgreeJudge(ctx context.Context, id int32, uuid string) (bool, error)
 	GetTalkCollectJudge(ctx context.Context, id int32, uuid string) (bool, error)
 	GetTalkSearch(ctx context.Context, page int32, search, time string) ([]*TalkSearch, int32, error)
+	GetUserTalkAgree(ctx context.Context, uuid string) (map[int32]bool, error)
+	GetUserTalkCollect(ctx context.Context, uuid string) (map[int32]bool, error)
 	GetTalkAuth(ctx context.Context, id int32) (int32, error)
+	GetCollectionsIdFromTalkCollect(ctx context.Context, id int32) (int32, error)
 
 	CreateTalkDraft(ctx context.Context, uuid string) (int32, error)
 	CreateTalkFolder(ctx context.Context, id int32, uuid string) error
@@ -38,23 +42,39 @@ type TalkRepo interface {
 	ReduceCreationUserTalk(ctx context.Context, auth int32, uuid string) error
 
 	SetTalkAgree(ctx context.Context, id int32, uuid string) error
+	SetUserTalkAgree(ctx context.Context, id int32, userUuid string) error
 	SetTalkView(ctx context.Context, id int32, uuid string) error
 	SetTalkViewToCache(ctx context.Context, id int32, uuid string) error
 	SetTalkAgreeToCache(ctx context.Context, id int32, uuid, userUuid string) error
 	SetTalkCollect(ctx context.Context, id int32, uuid string) error
 	SetTalkUserCollect(ctx context.Context, id, collectionsId int32, userUuid string) error
-	SetTalkCollectToCache(ctx context.Context, id int32, uuid, userUuid string) error
+
+	SetTalkCollectToCache(ctx context.Context, id, collectionsId int32, uuid, userUuid string) error
+	SetUserTalkAgreeToCache(ctx context.Context, id int32, userUuid string) error
+	SetUserTalkCollectToCache(ctx context.Context, id int32, userUuid string) error
+	SetCollectionsTalkCollect(ctx context.Context, id, collectionsId int32, userUuid string) error
+	SetCollectionTalk(ctx context.Context, collectionsId int32, userUuid string) error
+	SetUserTalkCollect(ctx context.Context, id int32, userUuid string) error
+	SetCreationUserCollect(ctx context.Context, userUuid string) error
 	CancelTalkAgree(ctx context.Context, id int32, uuid string) error
+	CancelUserTalkAgree(ctx context.Context, id int32, userUuid string) error
 	CancelTalkAgreeFromCache(ctx context.Context, id int32, uuid, userUuid string) error
 	CancelTalkUserCollect(ctx context.Context, id int32, userUuid string) error
 	CancelTalkCollect(ctx context.Context, id int32, uuid string) error
-	CancelTalkCollectFromCache(ctx context.Context, id int32, uuid, userUuid string) error
+	CancelTalkCollectFromCache(ctx context.Context, id, collectionsId int32, uuid, userUuid string) error
+	CancelUserTalkAgreeFromCache(ctx context.Context, id int32, userUuid string) error
+	CancelUserTalkCollectFromCache(ctx context.Context, id int32, userUuid string) error
+	CancelCollectionsTalkCollect(ctx context.Context, id int32, userUuid string) error
+	CancelUserTalkCollect(ctx context.Context, id int32, userUuid string) error
+	CancelCollectionTalk(ctx context.Context, collectionsId int32, userUuid string) error
+	ReduceCreationUserCollect(ctx context.Context, userUuid string) error
 
 	SendTalk(ctx context.Context, id int32, uuid string) (*TalkDraft, error)
 	SendReviewToMq(ctx context.Context, review *TalkReview) error
 	SendTalkToMq(ctx context.Context, talk *Talk, mode string) error
-	SendTalkStatisticToMq(ctx context.Context, uuid, mode string) error
+	SendTalkStatisticToMq(ctx context.Context, uuid, userUuid, mode string) error
 	SendScoreToMq(ctx context.Context, score int32, uuid, mode string) error
+	SendStatisticToMq(ctx context.Context, id, collectionsId int32, uuid, userUuid, mode string) error
 
 	DeleteTalk(ctx context.Context, id int32, uuid string) error
 	DeleteTalkDraft(ctx context.Context, id int32, uuid string) error
@@ -168,6 +188,22 @@ func (r *TalkUseCase) GetTalkSearch(ctx context.Context, page int32, search, tim
 		return nil, 0, v1.ErrorGetTalkSearchFailed("get talk search failed: %s", err.Error())
 	}
 	return talkList, total, nil
+}
+
+func (r *TalkUseCase) GetUserTalkAgree(ctx context.Context, uuid string) (map[int32]bool, error) {
+	agreeMap, err := r.repo.GetUserTalkAgree(ctx, uuid)
+	if err != nil {
+		return nil, v1.ErrorGetTalkAgreeFailed("get user talk agree failed: %s", err.Error())
+	}
+	return agreeMap, nil
+}
+
+func (r *TalkUseCase) GetUserTalkCollect(ctx context.Context, uuid string) (map[int32]bool, error) {
+	collectMap, err := r.repo.GetUserTalkCollect(ctx, uuid)
+	if err != nil {
+		return nil, v1.ErrorGetTalkAgreeFailed("get user talk collect failed: %s", err.Error())
+	}
+	return collectMap, nil
 }
 
 func (r *TalkUseCase) CreateTalkDraft(ctx context.Context, uuid string) (int32, error) {
@@ -376,6 +412,10 @@ func (r *TalkUseCase) DeleteTalkCacheAndSearch(ctx context.Context, id int32, uu
 			return v1.ErrorDeleteTalkFailed("freeze talk cos failed: %s", err.Error())
 		}
 
+		if auth == 2 {
+			return nil
+		}
+
 		err = r.repo.DeleteTalkSearch(ctx, id, uuid)
 		if err != nil {
 			return v1.ErrorDeleteTalkFailed("delete talk search failed: %s", err.Error())
@@ -385,8 +425,31 @@ func (r *TalkUseCase) DeleteTalkCacheAndSearch(ctx context.Context, id int32, uu
 }
 
 func (r *TalkUseCase) SetTalkAgree(ctx context.Context, id int32, uuid, userUuid string) error {
+	g, _ := errgroup.WithContext(ctx)
+	g.Go(r.re.GroupRecover(ctx, func(ctx context.Context) error {
+		err := r.repo.SetUserTalkAgreeToCache(ctx, id, userUuid)
+		if err != nil {
+			return v1.ErrorSetAgreeFailed("set user talk agree to cache failed: %s", err.Error())
+		}
+		return nil
+	}))
+	g.Go(r.re.GroupRecover(ctx, func(ctx context.Context) error {
+		err := r.repo.SendStatisticToMq(ctx, id, 0, uuid, userUuid, "set_talk_agree_db_and_cache")
+		if err != nil {
+			return v1.ErrorSetAgreeFailed("set talk agree to mq failed: %s", err.Error())
+		}
+		return nil
+	}))
+	return g.Wait()
+}
+
+func (r *TalkUseCase) SetTalkAgreeDbAndCache(ctx context.Context, id int32, uuid, userUuid string) error {
 	return r.tm.ExecTx(ctx, func(ctx context.Context) error {
 		err := r.repo.SetTalkAgree(ctx, id, uuid)
+		if err != nil {
+			return v1.ErrorSetAgreeFailed("set talk agree failed: %s", err.Error())
+		}
+		err = r.repo.SetUserTalkAgree(ctx, id, userUuid)
 		if err != nil {
 			return v1.ErrorSetAgreeFailed("set talk agree failed: %s", err.Error())
 		}
@@ -394,17 +457,44 @@ func (r *TalkUseCase) SetTalkAgree(ctx context.Context, id int32, uuid, userUuid
 		if err != nil {
 			return v1.ErrorSetAgreeFailed("set talk agree to cache failed: %s", err.Error())
 		}
-		err = r.repo.SendTalkStatisticToMq(ctx, uuid, "agree")
+		err = r.repo.SendTalkStatisticToMq(ctx, uuid, userUuid, "agree")
 		if err != nil {
 			return v1.ErrorSetAgreeFailed("set talk agree to mq failed: %s", err.Error())
+		}
+		err = r.repo.SendScoreToMq(ctx, 2, uuid, "add_score")
+		if err != nil {
+			return v1.ErrorSetAgreeFailed("send 2 score to mq failed: %s", err.Error())
 		}
 		return nil
 	})
 }
 
 func (r *TalkUseCase) CancelTalkAgree(ctx context.Context, id int32, uuid, userUuid string) error {
+	g, _ := errgroup.WithContext(ctx)
+	g.Go(r.re.GroupRecover(ctx, func(ctx context.Context) error {
+		err := r.repo.CancelUserTalkAgreeFromCache(ctx, id, userUuid)
+		if err != nil {
+			return v1.ErrorSetAgreeFailed("cancel user talk agree from cache failed: %s", err.Error())
+		}
+		return nil
+	}))
+	g.Go(r.re.GroupRecover(ctx, func(ctx context.Context) error {
+		err := r.repo.SendStatisticToMq(ctx, id, 0, uuid, userUuid, "cancel_talk_agree_db_and_cache")
+		if err != nil {
+			return v1.ErrorSetAgreeFailed("cancel talk agree to mq failed: %s", err.Error())
+		}
+		return nil
+	}))
+	return g.Wait()
+}
+
+func (r *TalkUseCase) CancelTalkAgreeDbAndCache(ctx context.Context, id int32, uuid, userUuid string) error {
 	return r.tm.ExecTx(ctx, func(ctx context.Context) error {
 		err := r.repo.CancelTalkAgree(ctx, id, uuid)
+		if err != nil {
+			return v1.ErrorCancelAgreeFailed("cancel talk agree failed: %s", err.Error())
+		}
+		err = r.repo.CancelUserTalkAgree(ctx, id, userUuid)
 		if err != nil {
 			return v1.ErrorCancelAgreeFailed("cancel talk agree failed: %s", err.Error())
 		}
@@ -412,15 +502,23 @@ func (r *TalkUseCase) CancelTalkAgree(ctx context.Context, id int32, uuid, userU
 		if err != nil {
 			return v1.ErrorCancelAgreeFailed("cancel talk agree from cache failed: %s", err.Error())
 		}
-		err = r.repo.SendTalkStatisticToMq(ctx, uuid, "agree_cancel")
+		err = r.repo.SendTalkStatisticToMq(ctx, uuid, userUuid, "agree_cancel")
 		if err != nil {
-			return v1.ErrorCancelAgreeFailed("set talk agree to mq failed: %s", err.Error())
+			return v1.ErrorCancelAgreeFailed("cancel talk agree to mq failed: %s", err.Error())
 		}
 		return nil
 	})
 }
 
 func (r *TalkUseCase) SetTalkView(ctx context.Context, id int32, uuid string) error {
+	err := r.repo.SendStatisticToMq(ctx, id, 0, uuid, "", "set_talk_view_db_and_cache")
+	if err != nil {
+		return v1.ErrorSetViewFailed("set talk view failed: %s", err.Error())
+	}
+	return nil
+}
+
+func (r *TalkUseCase) SetTalkViewDbAndCache(ctx context.Context, id int32, uuid string) error {
 	return r.tm.ExecTx(ctx, func(ctx context.Context) error {
 		err := r.repo.SetTalkView(ctx, id, uuid)
 		if err != nil {
@@ -430,53 +528,127 @@ func (r *TalkUseCase) SetTalkView(ctx context.Context, id int32, uuid string) er
 		if err != nil {
 			return v1.ErrorSetViewFailed("set talk view to cache failed: %s", err.Error())
 		}
-		err = r.repo.SendTalkStatisticToMq(ctx, uuid, "view")
+		err = r.repo.SendTalkStatisticToMq(ctx, uuid, "", "view")
 		if err != nil {
 			return v1.ErrorSetViewFailed("set talk view to mq failed: %s", err.Error())
+		}
+		err = r.repo.SendScoreToMq(ctx, 1, uuid, "add_score")
+		if err != nil {
+			return v1.ErrorSetViewFailed("send 1 score to mq failed: %s", err.Error())
 		}
 		return nil
 	})
 }
 
 func (r *TalkUseCase) SetTalkCollect(ctx context.Context, id, collectionsId int32, uuid, userUuid string) error {
+	g, _ := errgroup.WithContext(ctx)
+	g.Go(r.re.GroupRecover(ctx, func(ctx context.Context) error {
+		err := r.repo.SetUserTalkCollectToCache(ctx, id, userUuid)
+		if err != nil {
+			return v1.ErrorSetAgreeFailed("set user talk collect to cache failed: %s", err.Error())
+		}
+		return nil
+	}))
+	g.Go(r.re.GroupRecover(ctx, func(ctx context.Context) error {
+		err := r.repo.SendStatisticToMq(ctx, id, collectionsId, uuid, userUuid, "set_talk_collect_db_and_cache")
+		if err != nil {
+			return v1.ErrorSetAgreeFailed("set talk collect to mq failed: %s", err.Error())
+		}
+		return nil
+	}))
+	return g.Wait()
+}
+
+func (r *TalkUseCase) SetTalkCollectDbAndCache(ctx context.Context, id, collectionsId int32, uuid, userUuid string) error {
 	return r.tm.ExecTx(ctx, func(ctx context.Context) error {
-		err := r.repo.SetTalkUserCollect(ctx, id, collectionsId, userUuid)
+		err := r.repo.SetCollectionsTalkCollect(ctx, id, collectionsId, userUuid)
 		if err != nil {
 			return v1.ErrorSetCollectFailed("set talk user collect failed: %s", err.Error())
+		}
+		err = r.repo.SetCollectionTalk(ctx, collectionsId, userUuid)
+		if err != nil {
+			return v1.ErrorSetCollectFailed("set talk collection collect failed: %s", err.Error())
+		}
+		err = r.repo.SetUserTalkCollect(ctx, id, userUuid)
+		if err != nil {
+			return v1.ErrorSetCollectFailed("set user talk collect failed: %s", err.Error())
+		}
+		err = r.repo.SetCreationUserCollect(ctx, userUuid)
+		if err != nil {
+			return v1.ErrorSetCollectFailed("set creation user collect failed: %s", err.Error())
 		}
 		err = r.repo.SetTalkCollect(ctx, id, uuid)
 		if err != nil {
 			return v1.ErrorSetCollectFailed("set talk collect failed: %s", err.Error())
 		}
-		err = r.repo.SetTalkCollectToCache(ctx, id, uuid, userUuid)
+		err = r.repo.SetTalkCollectToCache(ctx, id, collectionsId, uuid, userUuid)
 		if err != nil {
 			return v1.ErrorSetCollectFailed("set talk collect to cache failed: %s", err.Error())
 		}
-		err = r.repo.SendTalkStatisticToMq(ctx, uuid, "collect")
+		err = r.repo.SendTalkStatisticToMq(ctx, uuid, "", "collect")
 		if err != nil {
 			return v1.ErrorSetCollectFailed("set talk collect to mq failed: %s", err.Error())
+		}
+		err = r.repo.SendScoreToMq(ctx, 2, uuid, "add_score")
+		if err != nil {
+			return v1.ErrorSetViewFailed("send 1 score to mq failed: %s", err.Error())
 		}
 		return nil
 	})
 }
 
 func (r *TalkUseCase) CancelTalkCollect(ctx context.Context, id int32, uuid, userUuid string) error {
-	return r.tm.ExecTx(ctx, func(ctx context.Context) error {
-		err := r.repo.CancelTalkUserCollect(ctx, id, userUuid)
+	g, _ := errgroup.WithContext(ctx)
+	g.Go(r.re.GroupRecover(ctx, func(ctx context.Context) error {
+		err := r.repo.CancelUserTalkCollectFromCache(ctx, id, userUuid)
 		if err != nil {
-			return v1.ErrorCancelCollectFailed("cancel talk user collect failed: %s", err.Error())
+			return v1.ErrorSetAgreeFailed("cancel user talk collect from cache failed: %s", err.Error())
+		}
+		return nil
+	}))
+	g.Go(r.re.GroupRecover(ctx, func(ctx context.Context) error {
+		err := r.repo.SendStatisticToMq(ctx, id, 0, uuid, userUuid, "cancel_talk_collect_db_and_cache")
+		if err != nil {
+			return v1.ErrorSetAgreeFailed("cancel talk collect to mq failed: %s", err.Error())
+		}
+		return nil
+	}))
+	return g.Wait()
+}
+
+func (r *TalkUseCase) CancelTalkCollectDbAndCache(ctx context.Context, id int32, uuid, userUuid string) error {
+	return r.tm.ExecTx(ctx, func(ctx context.Context) error {
+		collectionsId, err := r.repo.GetCollectionsIdFromTalkCollect(ctx, id)
+		if err != nil {
+			return v1.ErrorSetCollectFailed("get collections Id from collect failed: %s", err.Error())
+		}
+		err = r.repo.CancelCollectionsTalkCollect(ctx, id, userUuid)
+		if err != nil {
+			return v1.ErrorSetCollectFailed("cancel talk user collect failed: %s", err.Error())
+		}
+		err = r.repo.CancelUserTalkCollect(ctx, id, userUuid)
+		if err != nil {
+			return v1.ErrorSetCollectFailed("cancel user talk collect failed: %s", err.Error())
+		}
+		err = r.repo.CancelCollectionTalk(ctx, collectionsId, userUuid)
+		if err != nil {
+			return v1.ErrorSetCollectFailed("cancel talk collection collect failed: %s", err.Error())
+		}
+		err = r.repo.ReduceCreationUserCollect(ctx, userUuid)
+		if err != nil {
+			return v1.ErrorSetCollectFailed("reduce creation user collect failed: %s", err.Error())
 		}
 		err = r.repo.CancelTalkCollect(ctx, id, uuid)
 		if err != nil {
-			return v1.ErrorCancelCollectFailed("cancel talk collect failed: %s", err.Error())
+			return v1.ErrorSetCollectFailed("cancel talk collect failed: %s", err.Error())
 		}
-		err = r.repo.CancelTalkCollectFromCache(ctx, id, uuid, userUuid)
+		err = r.repo.CancelTalkCollectFromCache(ctx, id, collectionsId, uuid, userUuid)
 		if err != nil {
-			return v1.ErrorCancelCollectFailed("cancel talk collect to cache failed: %s", err.Error())
+			return v1.ErrorSetCollectFailed("cancel talk collect from cache failed: %s", err.Error())
 		}
-		err = r.repo.SendTalkStatisticToMq(ctx, uuid, "collect_cancel")
+		err = r.repo.SendTalkStatisticToMq(ctx, uuid, "", "collect_cancel")
 		if err != nil {
-			return v1.ErrorCancelCollectFailed("set talk collect to mq failed: %s", err.Error())
+			return v1.ErrorSetCollectFailed("cancel talk collect to mq failed: %s", err.Error())
 		}
 		return nil
 	})
